After discussion in #2, separated the wrapper into two wrappers: A high and low level one.
